+++ eventstore/mongodb/spring/blocking/src/test/java/org/occurrent/eventstore/mongodb/spring/blocking/SpringMongoEventStoreTest.java
@@ -1272,7 +1272,7 @@ public class SpringMongoEventStoreTest {
                 persist("name1", nameDefined);
 
                 // Then
-                Stream<CloudEvent> events = eventStore.all(EventStoreQueries.SortBy.NATURAL_ASC);
+                Stream<CloudEvent> events = eventStore.all(SortBy.natural(ASCENDING));
                 assertThat(deserialize(events)).containsExactly(nameWasChanged1, nameWasChanged2, nameDefined);
             }
 
@@ -1290,7 +1290,7 @@ public class SpringMongoEventStoreTest {
                 persist("name1", nameDefined);
 
                 // Then
-                Stream<CloudEvent> events = eventStore.all(EventStoreQueries.SortBy.NATURAL_DESC);
+                Stream<CloudEvent> events = eventStore.all(SortBy.natural(DESCENDING));
                 assertThat(deserialize(events)).containsExactly(nameDefined, nameWasChanged2, nameWasChanged1);
             }
 
@@ -1308,7 +1308,7 @@ public class SpringMongoEventStoreTest {
                 persist("name1", nameDefined);
 
                 // Then
-                Stream<CloudEvent> events = eventStore.all(EventStoreQueries.SortBy.TIME_ASC);
+                Stream<CloudEvent> events = eventStore.all(SortBy.time(ASCENDING));
                 assertThat(deserialize(events)).containsExactly(nameWasChanged1, nameDefined, nameWasChanged2);
             }
 
@@ -1326,7 +1326,7 @@ public class SpringMongoEventStoreTest {
                 persist("name1", nameDefined);
 
                 // Then
-                Stream<CloudEvent> events = eventStore.all(EventStoreQueries.SortBy.TIME_DESC);
+                Stream<CloudEvent> events = eventStore.all(SortBy.time(DESCENDING));
                 assertThat(deserialize(events)).containsExactly(nameDefined, nameWasChanged2, nameWasChanged1);
             }
         }