+++ subscription/util/blocking/catchup-subscription/src/main/java/org/occurrent/subscription/blocking/durable/catchup/CatchupSubscriptionModel.java
@@ -37,7 +37,7 @@ import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 import static org.occurrent.condition.Condition.gt;
-import static org.occurrent.eventstore.api.blocking.EventStoreQueries.SortBy.TIME_ASC;
+import static org.occurrent.eventstore.api.blocking.SortBy.time(ASCENDING);
 import static org.occurrent.filter.Filter.time;
 import static org.occurrent.functionalsupport.internal.FunctionalSupport.takeWhile;
 import static org.occurrent.time.internal.RFC3339.RFC_3339_DATE_TIME_FORMATTER;