+++ changelog.md
@@ -1,6 +1,7 @@
 ## Changelog next version
 
 * Only log with "warn" when subscription is restarted due to "ChangeStreamHistoryLost".
+* InMemoryEventStore now sorts queries by insertion order by default (before "time" was used) 
 
 ## Changelog 0.7.4 (2012-02-13)